switch to using verbose rather than summary flags
